SQL_Execution_New/requirements.txt
Flask
Flask-CORS
mysql-connector-python
trino
pyspark
requests
pandas
numpy